Chat service implemented

+++ event-service/src/main/java/com/waves/eventservice/model/Dto/ChatUser.java
@@ -11,9 +11,9 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ChatUser {
 
-    private Integer userId;
+    private Long userId;
     private String fullName;
     private String emailId;
-    private Integer eventId;
+    private Long eventId;
     private String eventName;
 }
